fix: Now properly waits after message completion, and started on data storage
